melhorias no layout

+++ poo-library-app-web/src/main/webapp/index.jsp
@@ -2,6 +2,8 @@
 <html ng-app="app">
 <head>
 <meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta name="author" content="José Nascimento &lt;joseaugustodearaujonascimento@gmail.com&gt;">
 <script src="assets/angular/angular.min.js" type="text/javascript"></script>
 <script type="text/javascript">
   var __f = function() {
@@ -20,6 +22,9 @@
       } ];
 
       $scope.delete = function(id) {
+
+        if (!confirm('Tem certeza?'))
+          return;
 
         $http({
           method : 'DELETE',
@@ -70,6 +75,9 @@
 .data-table th {
   background-color: lightgray;
 }
+input {
+    width: 100%;
+}
 </style>
 </head>
 <body>
@@ -88,8 +96,10 @@
       </tfoot>
       <tbody>
         <tr ng-repeat="user in users">
-          <td>\#{{user.id}}-{{user.nome}}</td>
-          <td>{{user.cpf}}</td>
+          <td>
+            \#{{user.id}}-{{user.nome}}</td>
+          <td>
+            {{user.cpf}}</td>
           <td>
             <button ng-click="delete(user.id)">delete</button>
           </td>
@@ -102,12 +112,12 @@
       method="post" ng-submit="save()">
       <div>
         <label> Nome: <input name="nome" type="text"
-          ng-model="data.nome">
+          ng-model="data.nome" minlength="10" required>
         </label>
       </div>
       <div>
-        <label> CPF: <input name="cpf" type="text"
-          ng-model="data.cpf">
+        <label> CPF: <input name="cpf" type="number"
+          ng-model="data.cpf" maxlength="11" minlength="11" required>
         </label>
       </div>
       <button class="usuario-create" type="submit">Salvar</button>
